fix: resolve macOS build issues and add connection retry logic
Build Fixes:
- Use dispatch.MainQueue().DispatchAsync() instead of objc.DispatchAsync
- Add baseLogger field to UI for audio.New() compatibility
- Add bytesToInt16() helper for audio chunk conversion
- Add GetFilePath() method to Config struct
- Fix uint vs uint64 type for foundation.Range
- Add static keyword to CGO functions to prevent duplicate symbols
- Fix nil pointer by deferring calibration handler setup until after window creation

Client Improvements:
- Add connection retry with exponential backoff (1s → 30s max)
- Client no longer gives up after first connection failure
- Properly handles server not being ready at startup

All builds pass on macOS with DarwinKit v0.5.0.

+++ client/internal/ui/window.go
@@ -92,7 +92,7 @@ func (w *Window) SetText(text string) {
 	// Scroll to bottom to show latest text
 	length := len(text)
 	if length > 0 {
-		w.textView.ScrollRangeToVisible(foundation.Range{Location: uint(length), Length: 0})
+		w.textView.ScrollRangeToVisible(foundation.Range{Location: uint64(length), Length: 0})
 	}
 }
